chatbot_rag/requirements.txt
flask==2.3.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pandas==2.0.3
numpy==1.24.3
requests==2.31.0